+++ lib/wechat/callback.rb
@@ -7,6 +7,9 @@ require 'wechat/callback/xml_document'
 require 'wechat/callback/message_decryption'
 require 'wechat/callback/message_encryption'
 
+require 'wechat/callback/random_byte_array'
+require 'wechat/callback/secure_message'
+
 module Wechat
   module Callback
     # Your code goes here...