+++ lib/wechat/callback/random_byte_array.rb
@@ -1,0 +1,9 @@
+class Wechat::Callback::RandomByteArray
+
+  LENGTH = 16
+
+  def self.create(length)
+    [*0..255].sample length
+  end
+
+end